Add support for releasing charms with resources
For charms that have file resources, like prometheus-k8s, the
upload-charm action will bind the latest matching file resource,
if any is found. This fits well the idea of publishing charms to
an edge, but will need revisiting when releasing channels to
other risk-levels like beta, where the promotion may occur with
charm and resource revisions that are not the latest.